requirements.txt

flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
Flask-Mail==0.9.1
eventlet==0.36.1
python-dotenv==1.0.1
supabase==2.14.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
openai==1.86.0
tavily-python==0.3.1
requests==2.32.3
httpx==0.28.1
aiohttp==3.11.18

playwright==1.50.0
playwright-mcp==0.1.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-whois==0.9.4
pandas==2.2.3
numpy==2.2.1
pydantic==2.10.3
cryptography==42.0.5
PyJWT==2.8.0
pyotp==2.9.0
docstring-parser==0.15
phonenumbers==8.13.40
pycountry==22.3.5
asyncio==3.4.3
aiofiles==24.1.0